+++ src/web/routes/application/flow-control/state-machine/predicates.test.js
@@ -1,0 +1,22 @@
+const test = require('tape')
+const { isPathAllowed } = require('./predicates')
+
+const paths = ['/first', '/second', '/third', '/fourth']
+
+test('isPathAllowed() should return true if path is before allowed in sequence', (t) => {
+  const result = isPathAllowed(paths, '/third', '/second')
+  t.equal(result, true, 'returns true if path is before allowed in sequence')
+  t.end()
+})
+
+test('isPathAllowed() should return true if path matches allowed', (t) => {
+  const result = isPathAllowed(paths, '/third', '/third')
+  t.equal(result, true, 'returns true if path matches allowed')
+  t.end()
+})
+
+test('isPathAllowed() should return false if path is after allowed in sequence', (t) => {
+  const result = isPathAllowed(paths, '/third', '/fourth')
+  t.equal(result, false, 'returns false if path is after allowed in sequence')
+  t.end()
+})
